+++ packages/modular-scripts/package.json
@@ -29,6 +29,7 @@
     "find-up": "^5.0.0",
     "fs-extra": "^9.0.1",
     "glob": "^7.1.6",
+    "http-proxy-agent": "^4.0.1",
     "mri": "^1.1.6",
     "node-fetch": "^2.6.1",
     "prompts": "^2.4.0",
