+++ packages/modular-scripts/src/cli.ts
@@ -1,6 +1,6 @@
 #!/usr/bin/env node
 
-import { JSONSchemaForNPMPackageJsonFiles as PackageJson, Dependency } from '@schemastore/package';
+import { JSONSchemaForNPMPackageJsonFiles as PackageJson } from '@schemastore/package';
 import mri from 'mri';
 import execa from 'execa';
 import * as fs from 'fs-extra';
@@ -14,8 +14,8 @@ import {
 import prompts from 'prompts';
 import resolveAsBin from 'resolve-as-bin';
 
-import fetch from './fetch';
 import getModularRoot from './getModularRoot';
+import preflight from './preflight';
 
 // Makes the script crash on unhandled rejections instead of silently
 // ignoring them. In the future, promise rejections that are not handled will
@@ -51,22 +51,13 @@ function execSync(
   });
 }
 
-function isYarnInstalled(): boolean {
-  try {
-    execa.sync('yarnpkg', ['-v']);
-    return true;
-  } catch (err) {
-    return false;
-  }
-}
-
 type PackageType = 'app' | 'view' | 'root'; // | 'package', the default
 
-export type ModularPackageJson = PackageJson & {
+type ModularPackageJson = PackageJson & {
   modular?: {
     type: PackageType;
   };
-}
+};
 
 function isModularType(dir: string, type: PackageType) {
   const packageJsonPath = path.join(dir, 'package.json');
@@ -336,44 +327,6 @@ async function build(directoryName: string, preserveModules?: boolean) {
     // ^ we do a dynamic import here to defer the module's initial side effects
     // till when it's actually needed (i.e. now)
     await build(directoryName, preserveModules);
-  }
-}
-
-interface NPMRegistryListing {
-  "dist-tags": Record<string, string>;
-}
-
-async function preflight() {
-  if (process.env.SKIP_PREFLIGHT_CHECK !== 'true') {
-    const { stdout: registry } = await execa('yarnpkg', ['config', 'get', 'registry']);
-
-    const url = String(new URL("/modular-scripts", registry));
-
-    const res = await fetch(url);
-    const modualarScriptsRegistry: NPMRegistryListing = await res.json();
-    const newVersion = modualarScriptsRegistry["dist-tags"]["latest"];
-
-    const { version }: ModularPackageJson = fs.readJSONSync(path.join(__dirname, "..", "package.json"));
-    if (newVersion != version) {
-      const { dependencies }: ModularPackageJson = fs.readJSONSync(path.join(getModularRoot(), "package.json"));
-
-      const update  = await prompts({
-        type: 'confirm',
-        name: 'value',
-        message: `Modular is out of date - would you like to update from ${version} to ${newVersion}`,
-        initial: true
-      });
-
-      if (update.value) {
-        await execa('yarnpkg', ['upgrade', 'modular-scripts']);
-      }
-    }
-  }
-
-  if (isYarnInstalled() === false) {
-    throw new Error(
-      'Please install `yarn` before attempting to run `modular-scripts`.',
-    );
   }
 }
